requirements.txt
-i https://pypi.org/simple
click==7.0
flask==1.0.2
gitdb2==2.0.5
gitpython==2.1.11
itsdangerous==1.1.0
jinja2==2.10
markupsafe==1.1.0
smmap2==2.0.5
werkzeug==0.14.1
